map_gen version 0.01
This is supper simple map genorator that genorates a simple map with one control point at 0,0,0 and one randomly placed asteroid. The position is randomly chosen from one of the six directions in 3d space. This means there are only six possible outcomes for the random map.  This will be made much better in latter verions.
Also I am pretty sure that the new lines will only work in a windows enviorment.